Rename consistent module to state
