+++ com.ibm.streamsx.topology/opt/python/packages/streamsx/topology/state.pyi
