+++ requirements.txt
@@ -1,4 +1,4 @@
 requests
-future
 dill
-enum34
+future
+enum34; python_version == '2.7'
